requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
dataclasses==0.8
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap-v5==1.0.4
django-filter==2.4.0
elasticsearch==7.13.4
fake-useragent==0.1.11
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
gunicorn==20.1.0
idna==3.2
idna-ssl==1.1.0
importlib-metadata==3.10.1
jsonschema==3.2.0
multidict==5.1.0
numpy==1.19.5
pandas==1.1.5
psycopg2==2.7.7
pycares==4.0.0
pycparser==2.20
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==1.2.4
pytz==2021.1
requests==2.26.0
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
-e git+https://github.com/twintproject/twint.git@e7c8a0c764f6879188e5c21e25fb6f1f856a7221#egg=twint
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
yarl==1.6.3
zipp==3.5.0
